+++ tests/CollectionTest.php
@@ -1,7 +1,7 @@
 <?php
 
 /**
- * Copyright 2016 Cloud Creativity Limited
+ * Copyright 2017 Cloud Creativity Limited
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,12 +16,21 @@
  * limitations under the License.
  */
 
-namespace CloudCreativity\Utils\Collection;
+namespace CloudCreativity\Utils\Collection\Tests;
 
 use ArrayObject;
+use CloudCreativity\Utils\Collection\Collection;
+use DateTime;
 use OutOfBoundsException;
+use PHPUnit\Framework\TestCase;
+use RuntimeException;
 use stdClass;
 
+/**
+ * Class CollectionTest
+ *
+ * @package CloudCreativity\Utils\Collection
+ */
 class CollectionTest extends TestCase
 {
 
@@ -30,7 +39,7 @@ class CollectionTest extends TestCase
         $expected = ['foo', 'bar'];
         $collection = new Collection($expected);
 
-        $this->assertSame($expected, $collection->toArray());
+        $this->assertSame($expected, $collection->all());
 
         return $collection;
     }
@@ -41,7 +50,7 @@ class CollectionTest extends TestCase
      */
     public function testTraversable(Collection $collection)
     {
-        $this->assertSame($collection->toArray(), iterator_to_array($collection));
+        $this->assertSame($collection->all(), iterator_to_array($collection));
     }
 
     /**
@@ -53,7 +62,7 @@ class CollectionTest extends TestCase
         $check = $collection->clear();
 
         $this->assertSame($collection, $check, 'Expecting `clear` to be chainable.');
-        $this->assertEmpty($collection->toArray());
+        $this->assertEmpty($collection->all());
     }
 
     public function testAdd()
@@ -64,11 +73,11 @@ class CollectionTest extends TestCase
         $check = $collection->add('foo');
 
         $this->assertSame($collection, $check, 'Expecting `add` to be chainable.');
-        $this->assertSame($expected, $collection->toArray());
-        $this->assertSame($expected, $collection->add('foo')->toArray());
+        $this->assertSame($expected, $collection->all());
+        $this->assertSame($expected, $collection->add('foo')->all());
 
         $expected[] = 'bar';
-        $this->assertSame($expected, $collection->add('bar')->toArray());
+        $this->assertSame($expected, $collection->add('bar')->all());
 
         return $collection;
     }
@@ -79,7 +88,7 @@ class CollectionTest extends TestCase
      */
     public function testAddIsStrict(Collection $collection)
     {
-        $expected = $collection->toArray();
+        $expected = $collection->all();
 
         array_push($expected, 10, (string) 10);
 
@@ -87,7 +96,7 @@ class CollectionTest extends TestCase
             ->add(10, true)
             ->add((string) 10, true);
 
-        $this->assertSame($expected, $collection->toArray());
+        $this->assertSame($expected, $collection->all());
     }
 
     public function testAddMany()
@@ -98,7 +107,7 @@ class CollectionTest extends TestCase
         $check = $collection->addMany(['foo', 'bar', 'bar']);
 
         $this->assertSame($collection, $check, 'Expecting `addMany` to be chainable.');
-        $this->assertSame($expected, $collection->toArray());
+        $this->assertSame($expected, $collection->all());
 
         return $collection;
     }
@@ -109,11 +118,11 @@ class CollectionTest extends TestCase
      */
     public function testAddManyIsStrict(Collection $collection)
     {
-        $expected = $collection->toArray();
+        $expected = $collection->all();
         array_push($expected, 10, (string) 10);
         $collection->addMany([10, (string) 10], true);
 
-        $this->assertSame($expected, $collection->toArray());
+        $this->assertSame($expected, $collection->all());
     }
 
     public function testAny()
@@ -148,7 +157,7 @@ class CollectionTest extends TestCase
             return false;
         });
 
-        $this->assertSame($collection->toArray(), $actual);
+        $this->assertSame($collection->all(), $actual);
     }
 
     public function testChunk()
@@ -208,7 +217,7 @@ class CollectionTest extends TestCase
             return true;
         });
 
-        $this->assertSame($collection->toArray(), $actual);
+        $this->assertSame($collection->all(), $actual);
     }
 
     public function testFind()
@@ -243,7 +252,7 @@ class CollectionTest extends TestCase
             return false;
         });
 
-        $this->assertSame($collection->toArray(), $actual);
+        $this->assertSame($collection->all(), $actual);
     }
 
     public function testContains()
@@ -284,7 +293,7 @@ class CollectionTest extends TestCase
 
         $this->assertFalse($collection->equals($compare));
 
-        $compare = new Collection(array_reverse($collection->toArray()));
+        $compare = new Collection(array_reverse($collection->all()));
 
         $this->assertFalse($collection->equals($compare));
 
@@ -333,7 +342,7 @@ class CollectionTest extends TestCase
             return true;
         });
 
-        $this->assertSame($collection->toArray(), $actual);
+        $this->assertSame($collection->all(), $actual);
     }
 
     public function testFirst()
@@ -374,7 +383,7 @@ class CollectionTest extends TestCase
      */
     public function testIndexOfOutOfBounds(Collection $collection)
     {
-        $this->setExpectedException(OutOfBoundsException::class);
+        $this->expectException(OutOfBoundsException::class);
         $collection->indexOf('a', count($collection));
     }
 
@@ -385,10 +394,10 @@ class CollectionTest extends TestCase
         $check = $collection->insertAt(1, 'b');
 
         $this->assertSame($collection, $check, 'Expecting `insertAt` to be chainable.');
-        $this->assertSame(['a', 'b', 'a'], $collection->toArray());
+        $this->assertSame(['a', 'b', 'a'], $collection->all());
 
         $collection->insertAt(3, 'b');
-        $this->assertSame(['a', 'b', 'a', 'b'], $collection->toArray());
+        $this->assertSame(['a', 'b', 'a', 'b'], $collection->all());
 
         return $collection;
     }
@@ -399,8 +408,35 @@ class CollectionTest extends TestCase
      */
     public function testInsertAtOutOfBounds(Collection $collection)
     {
-        $this->setExpectedException(OutOfBoundsException::class);
+        $this->expectException(OutOfBoundsException::class);
         $collection->insertAt(count($collection) + 1, 'foobar');
+    }
+
+    public function testInvoke()
+    {
+        $collection = new Collection([
+            $a = new DateTime(),
+            $b = new DateTime('-1 day'),
+            null
+        ]);
+
+        $expected = [$a->format('c'), $b->format('c'), null];
+        $this->assertSame($expected, $collection->invoke('format', 'c')->all());
+    }
+
+    public function testInvokeInvalidMethod()
+    {
+        $this->expectException(RuntimeException::class);
+        $this->expectExceptionMessage('DateTime::foo');
+        $collection = new Collection([new DateTime()]);
+        $collection->invoke('foo');
+    }
+
+    public function testInvokeNonObject()
+    {
+        $this->expectException(RuntimeException::class);
+        $collection = new Collection(['foo']);
+        $collection->invoke('foo');
     }
 
     public function testItemAt()
@@ -439,6 +475,15 @@ class CollectionTest extends TestCase
         $this->assertFalse($collection->isEmpty());
     }
 
+    public function testIsNotEmpty()
+    {
+        $collection = new Collection();
+
+        $this->assertFalse($collection->isNotEmpty());
+        $collection->push('a');
+        $this->assertTrue($collection->isNotEmpty());
+    }
+
     public function testUnique()
     {
         $collection = new Collection(['a', 'b', 'a', null, 'b', 10, '10']);
@@ -448,11 +493,11 @@ class CollectionTest extends TestCase
 
         /** Without strict */
         $expected = ['a', 'b', null, 10];
-        $this->assertSame($expected, $unique->toArray());
+        $this->assertSame($expected, $unique->all());
 
         /** With strict */
         $expected = ['a', 'b', null, 10, '10'];
-        $this->assertSame($expected, $collection->unique(true)->toArray());
+        $this->assertSame($expected, $collection->unique(true)->all());
     }
 
     public function testMap()
@@ -466,7 +511,7 @@ class CollectionTest extends TestCase
 
         $this->assertInstanceOf(Collection::class, $mapped);
         $this->assertNotSame($collection, $mapped);
-        $this->assertSame($expected, $mapped->toArray());
+        $this->assertSame($expected, $mapped->all());
     }
 
     public function testPop()
@@ -474,10 +519,10 @@ class CollectionTest extends TestCase
         $collection = new Collection(['a', 'b']);
 
         $this->assertSame('b', $collection->pop());
-        $this->assertSame(['a'], $collection->toArray());
+        $this->assertSame(['a'], $collection->all());
 
         $collection->pop();
-        $this->assertEmpty($collection->toArray());
+        $this->assertEmpty($collection->all());
         $this->assertNull($collection->pop());
     }
 
@@ -516,7 +561,7 @@ class CollectionTest extends TestCase
 
         $this->assertInstanceOf(Collection::class, $rejected);
         $this->assertNotSame($collection, $rejected);
-        $this->assertSame([10], $rejected->toArray());
+        $this->assertSame([10], $rejected->all());
 
         return $collection;
     }
@@ -533,7 +578,7 @@ class CollectionTest extends TestCase
             $actual[$key] = $value;
         });
 
-        $this->assertSame($collection->toArray(), $actual);
+        $this->assertSame($collection->all(), $actual);
     }
 
     public function testRemove()
@@ -541,7 +586,7 @@ class CollectionTest extends TestCase
         $collection = new Collection(['a', 'b', 'a']);
 
         $this->assertSame($collection, $collection->remove('a'));
-        $this->assertSame(['b'], $collection->toArray());
+        $this->assertSame(['b'], $collection->all());
 
         return $collection;
     }
@@ -552,10 +597,10 @@ class CollectionTest extends TestCase
         $other = $collection->copy();
 
         $collection->remove(10);
-        $this->assertEmpty($collection->toArray());
+        $this->assertEmpty($collection->all());
 
         $other->remove(10, true);
-        $this->assertSame(['10'], $other->toArray());
+        $this->assertSame(['10'], $other->all());
     }
 
     public function testRemoveAt()
@@ -564,7 +609,7 @@ class CollectionTest extends TestCase
         $expected = ['a', 'c'];
 
         $this->assertSame($collection, $collection->removeAt(1), 'Expecting `removeAt` to be chainable.');
-        $this->assertSame($expected, $collection->toArray());
+        $this->assertSame($expected, $collection->all());
 
         return $collection;
     }
@@ -575,7 +620,7 @@ class CollectionTest extends TestCase
      */
     public function testRemoveAtOutOfBounds(Collection $collection)
     {
-        $this->setExpectedException(OutOfBoundsException::class);
+        $this->expectException(OutOfBoundsException::class);
         $collection->removeAt(count($collection));
     }
 
@@ -583,8 +628,8 @@ class CollectionTest extends TestCase
     {
         $collection = new Collection(['a', 'b', 'b', 'b', 'c']);
 
-        $this->assertSame(['a', 'c'], $collection->copy()->removeAt(1, 3)->toArray());
-        $this->assertSame(['a', 'b', 'c'], $collection->removeAt(2, 2)->toArray());
+        $this->assertSame(['a', 'c'], $collection->copy()->removeAt(1, 3)->all());
+        $this->assertSame(['a', 'b', 'c'], $collection->removeAt(2, 2)->all());
     }
 
     public function testRemoveMany()
@@ -594,10 +639,10 @@ class CollectionTest extends TestCase
         $check = $collection->removeMany(['b', '10']);
 
         $this->assertSame($collection, $check, 'Expecting `removeMany` to be chainable.');
-        $this->assertSame(['a', 'c'], $collection->toArray());
+        $this->assertSame(['a', 'c'], $collection->all());
 
         /** Strict */
-        $this->assertSame(['a', 'c', 10], $other->removeMany(['b', '10'], true)->toArray());
+        $this->assertSame(['a', 'c', 10], $other->removeMany(['b', '10'], true)->all());
 
         return $collection;
     }
@@ -610,7 +655,7 @@ class CollectionTest extends TestCase
 
         $this->assertInstanceOf(Collection::class, $reversed);
         $this->assertNotSame($collection, $reversed);
-        $this->assertSame($expected, $reversed->toArray());
+        $this->assertSame($expected, $reversed->all());
     }
 
     public function testSearch()
@@ -636,7 +681,7 @@ class CollectionTest extends TestCase
         $this->assertNull($collection->shift());
         $collection->pushMany(['a', 'b']);
         $this->assertSame('a', $collection->shift());
-        $this->assertSame(['b'], $collection->toArray());
+        $this->assertSame(['b'], $collection->all());
     }
 
     public function testSlice()
@@ -648,8 +693,8 @@ class CollectionTest extends TestCase
         $sliced = $collection->slice(1);
         $this->assertInstanceOf(Collection::class, $sliced);
         $this->assertNotSame($collection, $sliced);
-        $this->assertSame($expected, $sliced->toArray());
-        $this->assertSame($original, $collection->toArray());
+        $this->assertSame($expected, $sliced->all());
+        $this->assertSame($original, $collection->all());
 
         return $collection;
     }
@@ -658,7 +703,7 @@ class CollectionTest extends TestCase
     {
         $collection = new Collection(['a', 'b', 'c', 'd']);
 
-        $this->assertSame(['b', 'c'], $collection->slice(1, 2)->toArray());
+        $this->assertSame(['b', 'c'], $collection->slice(1, 2)->all());
     }
 
     public function testSort()
@@ -677,8 +722,8 @@ class CollectionTest extends TestCase
 
         $this->assertInstanceOf(Collection::class, $sorted);
         $this->assertNotSame($collection, $sorted);
-        $this->assertSame($expected, $sorted->toArray());
-        $this->assertSame($original, $collection->toArray());
+        $this->assertSame($expected, $sorted->all());
+        $this->assertSame($original, $collection->all());
     }
 
     public function testSync()
@@ -689,10 +734,9 @@ class CollectionTest extends TestCase
         };
 
         $actual = $collection->sync(4, $callback);
-        $this->assertEquals(new Collection([10, 20, 30, 40]), $actual);
+        $this->assertSame([10, 20, 30, 40], $actual->all());
         $this->assertNotSame($collection, $actual);
-
-        $this->assertEquals(new Collection([10, 20]), $collection->sync(2, $callback));
+        $this->assertSame([10, 20], $collection->sync(2, $callback)->all());
 
         /** Overflow with no callable */
         $extra = new Collection();
@@ -700,8 +744,8 @@ class CollectionTest extends TestCase
             $extra->push($item);
         });
 
-        $this->assertEquals(new Collection([1, 2]), $actual);
-        $this->assertEquals(new Collection([3, 4]), $extra);
+        $this->assertSame([1, 2], $actual->all());
+        $this->assertSame([3, 4], $extra->all());
 
         /** Callable and overflow */
         $extra = new Collection();
@@ -709,15 +753,20 @@ class CollectionTest extends TestCase
             $extra->push($item);
         });
 
-        $this->assertEquals(new Collection([10, 20]), $actual);
-        $this->assertEquals(new Collection([3, 4]), $extra);
+        $this->assertSame([10, 20], $actual->all());
+        $this->assertSame([3, 4], $extra->all());
 
         /** Size is less than current length */
-        $this->assertEquals(new Collection([
-            10, 20, 30, 40, 'foo', 'foo'
-        ]), $collection->sync(6, $callback, function () {
+        $this->assertSame([10, 20, 30, 40, 'foo', 'foo'], $collection->sync(6, $callback, function () {
             $this->fail('Overflow callback should not be invoked when there is no overflow.');
-        }));
+        })->all());
+    }
+
+    public function testToArray()
+    {
+        $collection = new Collection($expected = ['a', 'b', 'c']);
+
+        $this->assertSame($expected, $collection->toArray());
     }
 
     public function testUnshift()
@@ -725,7 +774,7 @@ class CollectionTest extends TestCase
         $collection = new Collection(['a', 'b']);
 
         $this->assertSame($collection, $collection->unshift('c'));
-        $this->assertSame(['c', 'a', 'b'], $collection->toArray());
+        $this->assertSame(['c', 'a', 'b'], $collection->all());
     }
 
     public function testUnshiftMany()
@@ -734,7 +783,7 @@ class CollectionTest extends TestCase
         $expected = ['c', 'd', 'e', 'a', 'b'];
 
         $this->assertSame($collection, $collection->unshiftMany(['c', 'd', 'e']));
-        $this->assertSame($expected, $collection->toArray());
+        $this->assertSame($expected, $collection->all());
     }
 
     public function testWithout()
@@ -746,8 +795,8 @@ class CollectionTest extends TestCase
 
         $this->assertInstanceOf(Collection::class, $without);
         $this->assertNotSame($collection, $without);
-        $this->assertSame($expected, $without->toArray());
-        $this->assertSame($original, $collection->toArray());
+        $this->assertSame($expected, $without->all());
+        $this->assertSame($original, $collection->all());
     }
 
     public function testWithoutValueNotContained()
@@ -755,16 +804,16 @@ class CollectionTest extends TestCase
         $expected = ['a', 'b'];
         $collection = new Collection($expected);
 
-        $this->assertEquals($collection->toArray(), $collection->without('c')->toArray());
+        $this->assertEquals($collection->all(), $collection->without('c')->all());
     }
 
     public function testWithoutUsingStrict()
     {
         $collection = new Collection([1, 2, 3]);
 
-        $this->assertSame([1, 3], $collection->without('2')->toArray());
-        $this->assertSame([1, 2, 3], $collection->without('2', true)->toArray());
-        $this->assertSame([1, 3], $collection->without(2, true)->toArray());
+        $this->assertSame([1, 3], $collection->without('2')->all());
+        $this->assertSame([1, 2, 3], $collection->without('2', true)->all());
+        $this->assertSame([1, 3], $collection->without(2, true)->all());
     }
 
     public function testCast()