+++ tests/TestIterator.php
@@ -1,7 +1,7 @@
 <?php
 
 /**
- * Copyright 2016 Cloud Creativity Limited
+ * Copyright 2017 Cloud Creativity Limited
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,10 +16,28 @@
  * limitations under the License.
  */
 
-namespace CloudCreativity\Utils\Collection;
+namespace CloudCreativity\Utils\Collection\Tests;
 
+use CloudCreativity\Utils\Collection\AbstractStandardIterator;
+
+/**
+ * Class TestIterator
+ *
+ * @package CloudCreativity\Utils\Collection
+ */
 class TestIterator extends AbstractStandardIterator
 {
+
+    /**
+     * TestIterator constructor.
+     *
+     * @param array $items
+     */
+    public function __construct(array $items = [])
+    {
+        parent::__construct();
+        $this->addMany($items);
+    }
 
     /**
      * @param array $items